chore: removed redundant files
